Cambios en la estructura interna del código
