+++ src/GUI/sede.py
@@ -7,7 +7,10 @@ Clase disponible:
 
 import tkinter as tk
 from tkinter import ttk, messagebox
-from src.code.funciones import generar_id_unico, agregar_sede_archivo, obtener_sedes
+from src.code.constantes import JSON_SEDE, COLUMNAS_TABLA_SEDE, PREFIJO_SEDE
+from src.code.funciones import generar_id_unico
+from src.code.sede_code import obtener_sedes, agregar_sede_archivo
+
 
 class SedeApp:
     """
@@ -64,10 +67,9 @@ class SedeApp:
         Crea la tabla de sedes.
         """
         # Crear tabla
-        columnas = ("Nombre", "Ubicación", "Estado", "Teléfono")
-        self.tabla = ttk.Treeview(self.tabla_frame, columns=columnas, show="headings")
+        self.tabla = ttk.Treeview(self.tabla_frame, columns=COLUMNAS_TABLA_SEDE, show="headings")
 
-        for col in columnas:
+        for col in COLUMNAS_TABLA_SEDE:
             self.tabla.heading(col, text=col)
             self.tabla.column(col, minwidth=0, width=100)
 
@@ -79,6 +81,7 @@ class SedeApp:
         """
         Carga las sedes en la tabla.
         """
+
         sedes = obtener_sedes()
 
         for sede in sedes:
@@ -92,23 +95,10 @@ class SedeApp:
         ubicacion = self.ubicacion_var.get()
         estado = self.estado_var.get()
         telefono = self.telefono_var.get()
-
-        if not nombre or not ubicacion or not estado or not telefono:
-            messagebox.showerror("Error", "Por favor complete todos los campos.")
+        if not self.validarcampos( nombre, ubicacion, estado, telefono):
             return
-
-        try:
-            telefono = int(telefono)
-        except ValueError:
-            messagebox.showerror("Error", "El número de teléfono debe ser numérico.")
-            return
-
-        if len(str(telefono)) != 8:
-            messagebox.showerror("Error", "El número de teléfono debe tener 8 dígitos.")
-            return
-
         # Generar ID único para la sede
-        sede_id = generar_id_unico(self.prefijo)
+        sede_id = generar_id_unico(PREFIJO_SEDE)
 
         # Agregar nueva sede al archivo JSON
         agregar_sede_archivo(sede_id, nombre, ubicacion, estado, telefono)
@@ -125,3 +115,26 @@ class SedeApp:
         # Recargar la tabla de sedes
         self.tabla.delete(*self.tabla.get_children())
         self.cargar_sedes()
+
+    def validarcampos(self, nombre, ubicacion, estado, telefono):
+        sedes = obtener_sedes()
+        nombrevalido = True
+        for sede in sedes:
+            if sede["nombre"].upper() == nombre.upper():
+                nombrevalido = False
+        if not nombrevalido:
+            messagebox.showerror("Error", "Ya existe un sede con ese nombre")
+            return False
+        if not nombre or not ubicacion or not estado or not telefono:
+            messagebox.showerror("Error", "Por favor complete todos los campos.")
+            return False
+
+        try:
+            telefono = int(telefono)
+        except ValueError:
+            messagebox.showerror("Error", "El número de teléfono debe ser numérico.")
+            return False
+
+        if len(str(telefono)) != 8:
+            messagebox.showerror("Error", "El número de teléfono debe tener 8 dígitos.")
+            return False
