+++ src/main.py
@@ -1,10 +1,12 @@
 import tkinter as tk
-from src.GUI.ventanas import AppPrincipal
+from src.GUI.ventanas_admin import AppPrincipalAdmin
+
 
 def main():
     root = tk.Tk()
-    app = AppPrincipal(root)
+    app = AppPrincipalAdmin(root)
     root.mainloop()
+
 
 if __name__ == "__main__":
     main()
